training/app/requirements.txt
psycopg2-binary
pandas
scikit-learn
xgboost
